DAOL CMS 1.0.1.5

+++ modules/layout/layout.admin.controller.php
@@ -333,6 +333,9 @@
 				$ext = substr(strrchr($filename,'.'),1);
 				$filename = sprintf('%s.%s', md5($filename), $ext);
 			}
+			
+			// Check uploaded file
+			if(!checkUploadedFile($source['tmp_name'])) return false;
 
 			if(file_exists($path .'/'. $filename)) @unlink($path . $filename);
 			if(!move_uploaded_file($source['tmp_name'], $path . $filename )) return false;
@@ -575,7 +578,7 @@
 			// check upload
 			if(!Context::isUploaded()) exit();
 			$file = Context::get('file');
-			if(!is_uploaded_file($file['tmp_name'])) exit();
+			if(!is_uploaded_file($file['tmp_name']) || !checkUploadedFile($file['tmp_name'])) exit();
 			if(!preg_match('/\.(tar)$/i', $file['name'])) exit();
 
 			$layout_srl = Context::get('layout_srl');
@@ -767,7 +770,7 @@
 			$this->setTemplatePath($this->module_path.'tpl');
 			$this->setTemplateFile("after_upload_config_image.html");
 
-			if(!$img['tmp_name'] || !is_uploaded_file($img['tmp_name']))
+			if(!$img['tmp_name'] || !is_uploaded_file($img['tmp_name']) || !checkUploadedFile($img['tmp_name']))
 			{
 				Context::set('msg', Context::getLang('upload failed'));
 				return;
